remove authentication in huobi http, as i think of it now, it is not neccessary
