+++ exchange/huobi/http/interfaces.go
